+++ src/OpenSimBindings/Widgets/ModelEditorToolbar.cpp
@@ -116,13 +116,7 @@ private:
 
         if (ImGui::Button(ICON_FA_RECYCLE))
         {
-            ActionReloadOsimFromDisk(*m_Model);
-
-            // HACK (#594): purge the app-wide mesh cache so that any user edits to the underlying
-            // mesh files are immediately visible after reloading
-            //
-            // this is useful for users that are actively editing the meshes of the model file
-            App::upd().singleton<MeshCache>()->clear();
+            ActionReloadOsimFromDisk(*m_Model, *App::upd().singleton<MeshCache>());
         }
 
         if (!HasInputFileName(m_Model->getModel()))